Gevent thread pool for async_run
