+++ golem/task/tasksession.py
@@ -16,7 +16,7 @@ from golem.network.transport.session import MiddlemanSafeSession
 from golem.model import db
 from golem.model import Payment
 from golem.network.transport import tcpnetwork
-from golem.resource.client import AsyncRequest, async_run
+from golem.core.async import AsyncRequest, async_run
 from golem.resource.resource import decompress_dir
 from golem.task.taskbase import ComputeTaskDef, result_types, resource_types
 from golem.transactions.ethereum.ethereumpaymentskeeper import EthAccountInfo
